requirements.txt
CTkMessagebox
customtkinter
pillow
pyyaml